Drop Freenode #lisp channel
With the wave of channels moving from Freenode to Libera.chat, #lisp
and ##lisp have also moved. The new name of ##lisp is #lisp, and the
new name of #lisp is #commonlisp. Drop #commonlisp from our listing;
since it's no longer easily confused with #lisp, it's not directly
relevant to Scheme or to the Lisp family as a whole.